refactor: Исправлены замечания наставника
