+++ src/components/app/app.jsx
@@ -1,32 +1,39 @@
 import React from 'react';
 import {Switch, Route, BrowserRouter as Router} from 'react-router-dom';
 import {cardsPropTypes} from '../../prop-types';
-import MainScreen from '../main/main';
-import LoginScreen from '../login/login';
-import FavoritesScreen from '../favorites/favorites';
-import PropertyScreen from '../property/property';
-import NotFoundScreen from '../not-found/not-found';
+import MainScreen from '../main-screen/main-screen';
+import LoginScreen from '../login-screen/login-screen';
+import FavoritesScreen from '../favorites-screen/favorites-screen';
+import PropertyScreen from '../property-screen/property-screen';
+import NotFoundScreen from '../not-found-screen/not-found-screen';
+
+const routePathes = {
+  MAIN_SCREEN: `/`,
+  LOGIN_SCREEN: `/login`,
+  FAVORITES_SCREEN: `/favorites`,
+  PROPERTY_SCREEN: `/offer/:id`,
+};
 
 const App = ({cards}) => {
   return (
     <Router>
       <Switch>
 
-        <Route path="/" exact>
+        <Route path={routePathes.MAIN_SCREEN} exact>
           <MainScreen
             cards = {cards}
           />
         </Route>
 
-        <Route path="/login" exact>
+        <Route path={routePathes.LOGIN_SCREEN} exact>
           <LoginScreen />
         </Route>
 
-        <Route path="/favorites" exact>
+        <Route path={routePathes.FAVORITES_SCREEN} exact>
           <FavoritesScreen />
         </Route>
 
-        <Route path="/offer/:id" exact>
+        <Route path={routePathes.PROPERTY_SCREEN} exact>
           <PropertyScreen />
         </Route>
