+++ src/components/property-screen/property-screen.jsx
